Add attribute data-mw-tmh="" to the video/audio elements
The additional attribute will be used as selector in the follow-up
change Id1d7454ca31e804a94d45b43db146c650614ef7e.

The new selector restricts to elements created by extension
TimedMediaHandler. This prevents bad interaction with <audio> elements
created by extension Score.

The usage of the new attribute as selector in JavaScript is in a
separate change to allow to expire the cache of the rendered pages.

Bug: T295514
Change-Id: I744c8d5cad4d9b07b10b880f5acd87accabc8ec3