Properly honor $wgEnableTranscode
To fully disable transcodes, we need to set $wgEnableTranscode to false
and $wgEnabledAudioTranscodeSet to an empty array, which doesn't make
sense.

We do check $wgEnabledAudioTranscodeSet at the end of the function and
I don't see a purpose for checking it at the top, so this commit simply
removes it.

Bug: T374433
Change-Id: I5b4cc08e71108a8d9dac2aa9a2ae2decf5d9a0a7
(cherry picked from commit d9846dafb9c4c07a13c3566976f4c4d2ee56aebc)