Retool iOS MJPEG fallback video
Got confirmation that the iOS fallback support video was not working
through the HLS streaming on iOS 12, and can confirm locally on an
old iPad Air with iOS 12.5.7. This retools the fallback track as a
standalone QuickTime file, outside the HLS streaming, which should
be much more widely compatible with various versions of iOS still
out in the wild.

The HLS track experiment is being phased out as iOS 17.4 got full
WebM support, and iOS 18 is tightening up parameters on HLS.

Note that while QuickTime container is the same base format as ISO BMFF
and ".mp4", each flavor has different preferences on what's allowed.

Added QuickTime outputs for:

360p.mpeg4.mov: MPEG-4 Visual and MP3 audio at a good enough resolution
to be watchable IMO on a phone screen! 1 Mbit video

144p.mjpeg.mov: with lower-resolution Motion-JPEG, can be deployed before
we get confirmation on MPEG-4 Visual.

These are disabled by default, and would have to be enabled in config;
otherwise the ogv.js shim will be used on WebM playback on iOS 17.3
and earlier.

Bug: T363966
Change-Id: I56ccc2bd7d651a6104607420b25d7ecd22c86095