Avoid videojs menu's from flowing outside of the viewport
In vjs fullscreen dialog mode or in fullscreen, menu's had a risk of
overflowing the right side edge of the videoplayer, which in
fullscreen or in dialog mode would hide them outside the viewport.

By right side aligning these menu's to their parent, we avoid this.
This also helps simplify the alignment for rtl languages.

Bug: T379876
Change-Id: I41e0a027c5188e512849a3f14e13120f3a82b612