Allow more languages in the TimedText language dropdown
The language dropdown when creating a new TimedText file was
restricted to only languages that have a MessagesXxx.php file,
a restriction that doesn't make sense for subtitles. This
made languages defined in wgExtraLanguageNames not appear
in the dropdown, even though they are supported in the
video player itself when the pages exist.

Bug: T381934
Change-Id: Ib83ac52facd0eb8dcf4481fd46055405066be8d8