styles: Replace pixel font-size value for relative unit one
Relative values enable user browser setting overrides, one of the most
used accessibility features.
Not applying Codex design token yet as this would result in bumping up
the font size in various skins by 1px and 2px equivalent.
So staying with a hard-coded value for now as long as we don't have
designers and community involved in other direction.

Bug: T385857
Change-Id: Ie43520b4f12c41238b83d1b79ea9cf939bbb4b9a