Use a register callback instead of ExtensionFunctions
We used to require ExtensionFunctions because we were doing a lot of
setup in the function that required the config. Most of that is now
moved into other phases of the extention registration.

register callbacks are specifically to modify intial config and hook
setup. This makes the config safer/more predictable and allows us to
get rid of the globals as well.

Change-Id: Id3aa441cc583f6c9897e87e2feefe4948be1b1e3